+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/kyotakukeika/kyotakukeikakutodokede/KyotakuKeikakuTodokede.java
@@ -53,7 +53,7 @@ public class KyotakuKeikakuTodokede extends ParentModelBase<KyotakuKeikakuTodoke
      */
     public KyotakuKeikakuTodokede(HihokenshaNo 被保険者番号,
             FlexibleYearMonth 対象年月,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         requireNonNull(被保険者番号, UrSystemErrorMessages.値がnull.getReplacedMessage("被保険者番号"));
         requireNonNull(対象年月, UrSystemErrorMessages.値がnull.getReplacedMessage("対象年月"));
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
@@ -135,7 +135,7 @@ public class KyotakuKeikakuTodokede extends ParentModelBase<KyotakuKeikakuTodoke
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public int get履歴番号() {
         return entity.getRirekiNo();
     }
